% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-exportWithDate.R
\name{exportWithDate}
\alias{exportWithDate}
\title{exportWithDate}
\usage{
exportWithDate(x, filename)
}
\arguments{
\item{x}{Dataframe to be exported.}

\item{filename}{Filename}
}
\value{
Output in local
}
\description{
Export data frame based on stored filename alongside exporting date.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
